% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weitrix.R
\name{as_weitrix}
\alias{as_weitrix}
\title{Ensure data is a weighted matrix}
\usage{
as_weitrix(object, weights = NULL)
}
\description{
Ensure data is a weighted matrix. 
This is a SummarizedExperiment or subclass thereof with some metadata fields set. 
If it is ambiguous how to do this, produce an error.
}
