% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{components_bootstrap}
\alias{components_bootstrap}
\title{Experimental! Bootstrap resample columns of a weitrix then refit components of variation}
\usage{
components_bootstrap(comp, weitrix, max_iter = 5, verbose = TRUE)
}
\arguments{
\item{comp}{A Components object, as produced by weitrix_components.}

\item{weitrix}{The weitrix (or an object that can be converted to a weitrix with \code{as_weitrix}) from which comp was produced.}

\item{max_iter}{Maximum iterations when refitting components.}

\item{verbose}{Show messages about the progress of the iterations when refitting.}
}
\description{
A bootstrap version of the weitrix is created, and then components refitted by re-optimizing rows, cols, rows, cols, rows, cols, etc (up to \code{max_iter}).
}
\details{
This is experimental, and may be changed or removed in future.
}
