% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components_seq_screeplot}
\alias{components_seq_screeplot}
\title{Produce a scree plot}
\usage{
components_seq_screeplot(comp_seq, rand_seq = NULL)
}
\arguments{
\item{comp_seq}{A list of Components objects, as produced by \code{components_seq}.}

\item{rand_seq}{Optional. A Components object with a single component, or list of Components objects.}
}
\description{
Produce a scree plot
}
