% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components_varimax}
\alias{components_varimax}
\title{Seek meaningful components by varimax rotation}
\usage{
components_varimax(comp)
}
\arguments{
\item{comp}{A Components object.}
}
\description{
Varimax rotation rotates components in a decomposition so that
    each component has a few large loadings and the rest small.
By the grace of sparsity we hope that these components 
    are then individually interpretable.
}
