% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{counts_in_totals_proportions}
\alias{counts_in_totals_proportions}
\title{Weitrix of proportions of totals}
\usage{
counts_in_totals_proportions(counts, totals)
}
\arguments{
\item{counts}{A matrix (or DelayedArray) of counts.}

\item{totals}{A matrix (or DelayedArray) of totals, of the same dimensions as the counts matrix.}
}
\value{
A SummarizedExperiment object with metadata fields marking it as a weitrix.
}
\description{
Construct a weitrix of proportions by dividing a matrix of counts by a matrix of totals. Each proportion is given the total as its weight. The reasoning is that a proportion is an average of "count" 1s and "total-count" 0s. When values are averages, the variance is inversely proportional to the number of individual items being averaged.
}
\details{
A reasonable default formula is set in the metadata for use with \code{weitrix_calibrate_trend}. This uses two columns included in the row data. These are the total of totals ("total_reads"), and the expected variance if proportions are uniform ("per_read_var"). 

Where the total is zero, the weight will be zero and the proportion will be NA.

Counts and totals do not need to be integers, as long as the basic idea of variance being inversely proportional to the "total" holds.
}
