% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simwei}
\alias{simwei}
\title{Simulated weitrix dataset.}
\format{
A weitrix object.
}
\usage{
simwei
}
\description{
This is a small simulated weitrix used in examples.
There is one component of variation to be found, 
    plus Gaussian noise with variance inversely proportional to the weights.
}
\keyword{datasets}
