% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{weitrix_bootstrap}
\alias{weitrix_bootstrap}
\title{Bootstrap resample columns of a weitrix}
\usage{
weitrix_bootstrap(weitrix)
}
\arguments{
\item{weitrix}{A weitrix object, or an object that can be converted to a weitrix with \code{as_weitrix}.}
}
\description{
Bootstrap resample a weitrix. Where a column is sampled multiple times its weights are increased that many times, rather than producing multiple columns.
}
\details{
Note that this will not work with methods that count the residual degrees of freedom, such as estimating the residual standard deviation. The main applicaition is with linear model fitting by weighted least squares, and weitrix_components which is based on this. These will work correctly.
}
