% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{weitrix_calibrate}
\alias{weitrix_calibrate}
\title{Adjust weights based on given row dispersions}
\usage{
weitrix_calibrate(weitrix, dispersions)
}
\description{
For large numbers of samples this can be based directly on weitrix_dispersions. For small numbers of samples, when using limma, it should be based on a trend-line fitted to known co-variates of the dispersions estimated by weitrix_dispersions.
}
