% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{weitrix_dispersions}
\alias{weitrix_dispersions}
\title{Calculate row dispersions}
\usage{
weitrix_dispersions(weitrix, comp)
}
\arguments{
\item{weitrix}{A weitrix object, or an object that can be converted to a weitrix 
with \code{as_weitrix}.}

\item{comp}{A Components, an approximate matrix decomposition of 
the weitrix x values, for example created with \code{weitrix_components}.}
}
\description{
Calculate the dispersion of each row. For each observation, 
this value divided by the weight gives the observation's variance.
}
