% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_count.R
\name{weitrix_downweight_sparse}
\alias{weitrix_downweight_sparse}
\title{Downweight rows and cols with low effective numbers of observations}
\usage{
weitrix_downweight_sparse(weitrix, power = 2, order = 2, rows = TRUE,
  cols = TRUE)
}
\description{
This is an alternative to filtering out sparse rows and columns, generally to be used before weitrix_components. For each row, weights are scaled by the row's Hill number of order \code{order} as a fraction of the total possible, raised to the power \code{power}. The same is done for each column.
}
