% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_count.R
\name{weitrix_hill}
\alias{weitrix_hill}
\title{Calculate Hill numbers (effective number of observations) for rows or columns}
\usage{
weitrix_hill(weitrix, what = c("row", "col"), order = 2)
}
\arguments{
\item{weitrix}{A weitrix object.}

\item{what}{Calculate for rows ("row") (default) or columns ("col")?}

\item{order}{Order of the Hill numbers.}
}
\description{
Effective numbers of observations. order=0 produces count of non-zero weights. order=1 produces exp(entropy). order=2 produces the inverse Simpson index.
}
