% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{weitrix_randomize}
\alias{weitrix_randomize}
\title{Generate a random normally distributed version of a weitrix}
\usage{
weitrix_randomize(weitrix)
}
\arguments{
\item{weitrix}{A weitrix object, or an object that can be converted to a weitrix with \code{as_weitrix}.}
}
\description{
Values are generated with variance equal to 1/weight.
This can be used to see what R-squared would be achieved with purely random data,
and therefore an appropriate number of components to use.
This is known as Parallel Analysis.
}
\seealso{
\code{\link{components_seq_screeplot}}
}
