% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weitrix.R
\name{weitrix_weights}
\alias{weitrix_weights}
\alias{weitrix_weights<-}
\title{Get or set a weitrix object's "weights" matrix}
\usage{
weitrix_weights(weitrix)

weitrix_weights(x) <- value
}
\arguments{
\item{weitrix}{A weitrix object.}

\item{x}{The weitrix to modify.}

\item{value}{The new matrix.}
}
\value{
A matrix-like object such as a matrix or a DelayedArray.
}
\description{
Gets or sets the appropriate assay in the SummarizedExperiment object.
}
\examples{
weitrix_weights(simwei)

}
