% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weitrix.R
\name{weitrix_x}
\alias{weitrix_x}
\alias{weitrix_x<-}
\title{Get or set a weitrix object's "x" matrix}
\usage{
weitrix_x(weitrix)

weitrix_x(x) <- value
}
\arguments{
\item{weitrix}{A weitrix object.}

\item{x}{The weitrix to modify.}

\item{value}{The new matrix.}
}
\value{
A matrix-like object such as a matrix or a DelayedArray.
}
\description{
Gets or sets the appropriate assay in the SummarizedExperiment object.
}
\examples{
weitrix_x(simwei)

simwei2 <- simwei
weitrix_x(simwei2) <- weitrix_x(simwei2) * 2

}
