% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weitrix.R
\name{as_weitrix}
\alias{as_weitrix}
\title{Ensure data is a weighted matrix}
\usage{
as_weitrix(object, weights = NULL)
}
\description{
Ensure data is a weighted matrix or "weitrix". 
A weitrix is a SummarizedExperiment or subclass thereof with some metadata fields set. 
If it is ambiguous how to do this, produce an error.
}
\details{
Input can be a matrix or DelayedArray.

Input can be anything the limma package recognizing,
notably the EList class (for example as output by voom or vooma).

If weights are not present in "object" and not given with "weights",
they default for 0 for NA values and 1 for everything else.
}
