% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{components_seq_scree}
\alias{components_seq_scree}
\title{Proportion more variance explained by adding components one at a time}
\usage{
components_seq_scree(comp_seq)
}
\description{
Proportion more variance explained by adding components one at a time
}
