% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{matrix_long}
\alias{matrix_long}
\title{Convert a matrix to long form for ggplotting}
\usage{
matrix_long(
  matrix,
  row_info = NULL,
  col_info = NULL,
  varnames = c("name", "col")
)
}
\description{
Convert a matrix to long form for ggplotting
}
