% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{weitrix_calibrate_trend}
\alias{weitrix_calibrate_trend}
\title{Adjust weights by fitting a trend to estimated observed dispersions}
\usage{
weitrix_calibrate_trend(weitrix, comp, formula = NULL)
}
\description{
Adjust weights by fitting a trend to estimated observed dispersions
}
