% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{weitrix_components}
\alias{weitrix_components}
\title{Principal components of a weitrix}
\usage{
weitrix_components(weitrix, p = 2, design = NULL, max_iter = 100,
  tol = 1e-05, verbose = TRUE, initial = NULL)
}
\description{
Principal components of a weitrix
}
