% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{weitrix_confects}
\alias{weitrix_confects}
\title{Top confident effects based on one or more contrasts of a linear model for each row}
\usage{
weitrix_confects(
  weitrix,
  design,
  contrasts,
  effect = c("auto", "contrast", "sd", "cohen_f"),
  dispersion_est = c("ebayes_limma", "row", "none"),
  fdr = 0.05,
  step = NULL,
  full = FALSE
)
}
\description{
This function provides topconfects-style testing of a linear model contrast, as well as a multi-contrast extension to this method for F-tests with effect sizes.
}
