% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weitrix.R
\name{weitrix_elist}
\alias{weitrix_elist}
\title{Convert a weitrix object to a limma EList object}
\usage{
weitrix_elist(weitrix)
}
\arguments{
\item{weitrix}{A weitrix object.}
}
\description{
The resulting object can be used as input to \code{limma::lmFit}
for a limma analysis.
}
