% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{weitrix_randomize}
\alias{weitrix_randomize}
\title{Generate a random normally distributed version of a weitrix}
\usage{
weitrix_randomize(weitrix, BPPARAM = getAutoBPPARAM())
}
\description{
Weights are used to choose the standard deviation of values generated.
}
