% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weitrix.R
\name{weitrix_x}
\alias{weitrix_x}
\alias{weitrix_x<-}
\title{Get or set a weitrix object's "x" matrix}
\usage{
weitrix_x(weitrix)

weitrix_x(x) <- value
}
\arguments{
\item{weitrix}{A weitrix object.}

\item{x}{The weitrix to modify.}

\item{value}{The new matrix.}
}
\description{
Gets or sets the appropriate assay in the SummarizedExperiment object.
}
